Fix for DIRSERVER-1735 (When ads-pwdmaxage attribute is set to more then 2147483, then all user passwords are expirer forever.).
Thanks to Oldrich Novak for providing the bug report and the fix.

git-svn-id: https://svn.apache.org/repos/asf/directory/apacheds/trunk@1362460 13f79535-47bb-0310-9956-ffa450edef68

+++ core-api/src/main/java/org/apache/directory/server/core/api/authn/PasswordUtil.java
@@ -496,7 +496,7 @@ public class PasswordUtil
     {
         Date pwdChangeDate = DateUtils.getDate( pwdChangedZtime );
 
-        long time = pwdMaxAgeSec * 1000;
+        long time = ( ( long ) pwdMaxAgeSec ) * 1000L;
         time += pwdChangeDate.getTime();
 
         Date expiryDate = DateUtils.getDate( DateUtils.getGeneralizedTime( time ) );
